// api-gateway/src/auth/types/login-kafka-response.type.ts
export type LoginKafkaResponse = {
  user: {
    id: string;
    email: string;
    systemRole: string;
    status: string;
    emailVerifiedAt: string | Date | null;
    createdAt: string | Date;
  };

  tokens: {
    accessToken: string;
    accessTokenExpiresAt: string | Date;
    refreshToken: string;
    refreshTokenExpiresAt: string | Date;
  };
};